+++ assets/src/components/activities/vlab/VlabDelivery.tsx
@@ -4,20 +4,18 @@ import { ResetButtonConnected } from 'components/activities/common/delivery/rese
 import { SubmitButtonConnected } from 'components/activities/common/delivery/submit_button/SubmitButtonConnected';
 import { HintsDeliveryConnected } from 'components/activities/common/hints/delivery/HintsDeliveryConnected';
 import { StemDelivery } from 'components/activities/common/stem/delivery/StemDelivery';
-import {
-  DeliveryElement,
-  DeliveryElementProps,
-  DeliveryElementProvider,
-  useDeliveryElementContext,
-} from 'components/activities/DeliveryElement';
-import { MultiInputSchema } from 'components/activities/vlab/schema';
+import { DeliveryElement, DeliveryElementProps } from 'components/activities/DeliveryElement';
+import { DeliveryElementProvider, useDeliveryElementContext } from '../DeliveryElementProvider';
+import { VlabSchema } from 'components/activities/vlab/schema';
 import { Manifest, PartId } from 'components/activities/types';
-import { toSimpleText } from 'components/editing/utils';
+import { toSimpleText } from 'components/editing/slateUtils';
 import {
   activityDeliverySlice,
   ActivityDeliveryState,
   initializeState,
   isEvaluated,
+  listenForParentSurveySubmit,
+  listenForParentSurveyReset,
   PartInputs,
   resetAction,
 } from 'data/activities/DeliveryState';
@@ -32,16 +30,86 @@ import { configureStore } from 'state/store';
 export const MultiInputComponent: React.FC = () => {
   const {
     state: activityState,
+    context,
+    onSubmitActivity,
     onSaveActivity,
     onResetActivity,
     model,
-    sectionSlug,
-  } = useDeliveryElementContext<MultiInputSchema>();
+  } = useDeliveryElementContext<VlabSchema>();
+
   const uiState = useSelector((state: ActivityDeliveryState) => state);
   const [hintsShown, setHintsShown] = React.useState<PartId[]>([]);
   const dispatch = useDispatch();
 
+  const onVlabChange = () => {
+    // Get the selected flask XML and parse.
+
+    // TODO: this will break when more than one instance of Vlab is present on a page
+    const selectedFlaskXML = (
+      document.getElementById('vlab') as any
+    ).contentWindow.getSelectedItem();
+    const parser = new DOMParser();
+    const selectedFlask = parser.parseFromString(selectedFlaskXML, 'application/xml');
+
+    // Loop over the inputs, if an input is type vlabInput, update it's value based on XML.
+    if (!selectedFlask.querySelector('parsererror')) {
+      let value: any = 0;
+      model.inputs.forEach((input) => {
+        if (input.inputType === 'vlabvalue') {
+          // Move this mess to Utils?
+          // value = vlabValueFromXML(input, selectedFlask);
+          const param = input.parameter;
+          const volume = selectedFlask
+            .getElementsByTagName('flask')[0]
+            .getElementsByTagName('volume')[0].textContent as any;
+          const speciesList = Array.from(
+            selectedFlask.getElementsByTagName('flask')[0].getElementsByTagName('species'),
+          );
+          if (param === 'volume' || param === 'temp') {
+            value = selectedFlask
+              .getElementsByTagName('flask')[0]
+              .getElementsByTagName(param)[0].textContent;
+          } else if (param === 'pH') {
+            speciesList.forEach((species) => {
+              if (species.getElementsByTagName('id')[0].textContent === '1') {
+                const hPlusMolarity =
+                  (species.getElementsByTagName('moles') as any)[0].textContent / volume;
+                value = -1 * Math.log10(hPlusMolarity);
+              }
+            });
+          } else {
+            speciesList.forEach((species) => {
+              if (species.getElementsByTagName('id')[0].textContent === input.species) {
+                if (param === 'molarity') {
+                  value = (species.getElementsByTagName('moles') as any)[0].textContent / volume;
+                } else if (param === 'concentration') {
+                  value = (species.getElementsByTagName('mass') as any)[0].textContent / volume;
+                } else {
+                  value = (species.getElementsByTagName(param) as any)[0].textContent;
+                }
+              }
+            });
+          }
+          dispatch(
+            activityDeliverySlice.actions.setStudentInputForPart({
+              partId: input.partId,
+              studentInput: [value as any],
+            }),
+          );
+        }
+      });
+    }
+  };
+
+  const emptyPartInputs = model.inputs.reduce((acc: any, input: any) => {
+    acc[input.partId] = [''];
+    return acc;
+  }, {} as PartInputs);
+
   useEffect(() => {
+    listenForParentSurveySubmit(context.surveyId, dispatch, onSubmitActivity);
+    listenForParentSurveyReset(context.surveyId, dispatch, onResetActivity, emptyPartInputs);
+
     dispatch(
       initializeState(
         activityState,
@@ -53,8 +121,14 @@ export const MultiInputComponent: React.FC = () => {
               return acc;
             }, {} as PartInputs),
         }),
+        model,
+        context,
       ),
     );
+    window.addEventListener('message', onVlabChange);
+    return () => {
+      window.removeEventListener('message', onVlabChange);
+    };
   }, []);
 
   // First render initializes state
@@ -63,7 +137,7 @@ export const MultiInputComponent: React.FC = () => {
   }
 
   const toggleHints = (id: string) => {
-    const input = getByUnsafe(model.inputs, (x) => x.id === id);
+    const input = getByUnsafe((uiState.model as VlabSchema).inputs, (x) => x.id === id);
     setHintsShown((hintsShown) =>
       hintsShown.includes(input.partId)
         ? hintsShown.filter((id) => id !== input.partId)
@@ -72,7 +146,7 @@ export const MultiInputComponent: React.FC = () => {
   };
 
   const inputs = new Map(
-    model.inputs.map((input) => [
+    (uiState.model as VlabSchema).inputs.map((input) => [
       input.id,
       {
         input:
@@ -80,7 +154,7 @@ export const MultiInputComponent: React.FC = () => {
             ? {
                 id: input.id,
                 inputType: input.inputType,
-                options: model.choices
+                options: (uiState.model as VlabSchema).choices
                   .filter((c) => input.choiceIds.includes(c.id))
                   .map((choice) => ({
                     value: choice.id,
@@ -94,9 +168,9 @@ export const MultiInputComponent: React.FC = () => {
     ]),
   );
 
-  const onChange = (id: string, e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
-    const input = getByUnsafe(model.inputs, (x) => x.id === id);
-    const value = e.target.value;
+  const onChange = (id: string, value: string) => {
+    const input = getByUnsafe((uiState.model as VlabSchema).inputs, (x) => x.id === id);
+
     dispatch(
       activityDeliverySlice.actions.setStudentInputForPart({
         partId: input.partId,
@@ -113,12 +187,33 @@ export const MultiInputComponent: React.FC = () => {
     ]);
   };
 
+  const onVlabLoad = () => {
+    if (model.assignmentSource === 'builtIn') {
+      const assignment = model.assignmentPath;
+      (document.getElementById('vlab') as any).contentWindow.loadAssignment(assignment);
+    } else {
+      const assignmentJSON = {
+        assignment: JSON.parse(model.assignment),
+        configuration: JSON.parse(model.configuration),
+        reactions: JSON.parse(model.reactions),
+
+        // TODO: these attributes are not on the VLab model:
+        solutions: JSON.parse((model as any).solutions),
+        species: JSON.parse((model as any).species),
+        spectra: JSON.parse((model as any).spectra),
+      };
+      (document.getElementById('vlab') as any).contentWindow.loadAssignmentJSON(assignmentJSON);
+    }
+  };
+
   const writerContext = defaultWriterContext({
-    sectionSlug,
+    sectionSlug: context.sectionSlug,
+    bibParams: context.bibParams,
     inputRefContext: {
       toggleHints,
       onChange,
-      inputs,
+      // TODO: This 'as any' cast was necessary as the types do not align
+      inputs: inputs as any,
       disabled: isEvaluated(uiState),
     },
   });
@@ -126,20 +221,15 @@ export const MultiInputComponent: React.FC = () => {
   return (
     <div className="activity mc-activity">
       <div className="activity-content">
-        <StemDelivery className="form-inline" stem={model.stem} context={writerContext} />
+        <iframe id="vlab" className="vlab-holder" src="/vlab/vlab.html" onLoad={onVlabLoad} />
+        <StemDelivery
+          className="form-inline"
+          stem={(uiState.model as VlabSchema).stem}
+          context={writerContext}
+        />
         <GradedPointsConnected />
         <ResetButtonConnected
-          onReset={() =>
-            dispatch(
-              resetAction(
-                onResetActivity,
-                model.inputs.reduce((acc, input) => {
-                  acc[input.partId] = [''];
-                  return acc;
-                }, {} as PartInputs),
-              ),
-            )
-          }
+          onReset={() => dispatch(resetAction(onResetActivity, emptyPartInputs))}
         />
         <SubmitButtonConnected disabled={false} />
         {hintsShown.map((partId) => (
@@ -156,8 +246,8 @@ export const MultiInputComponent: React.FC = () => {
 };
 
 // Defines the web component, a simple wrapper over our React component above
-export class VlabDelivery extends DeliveryElement<MultiInputSchema> {
-  render(mountPoint: HTMLDivElement, props: DeliveryElementProps<MultiInputSchema>) {
+export class VlabDelivery extends DeliveryElement<VlabSchema> {
+  render(mountPoint: HTMLDivElement, props: DeliveryElementProps<VlabSchema>) {
     const store = configureStore({}, activityDeliverySlice.reducer);
     ReactDOM.render(
       <Provider store={store}>