+++ assets/src/components/activities/vlab/authoring-entry.ts
@@ -4,12 +4,12 @@ export { VlabAuthoring } from './VlabAuthoring';
 // Registers the creation function:
 import { Manifest, CreationContext } from '../types';
 import { registerCreationFunc } from '../creation';
-import { MultiInputSchema } from './schema';
+import { VlabSchema } from './schema';
 import { defaultModel } from './utils';
 // eslint-disable-next-line
 const manifest: Manifest = require('./manifest.json');
 
-function createFn(_context: CreationContext): Promise<MultiInputSchema> {
+function createFn(_context: CreationContext): Promise<VlabSchema> {
   return Promise.resolve(Object.assign({}, defaultModel()));
 }