+++ assets/src/components/activities/vlab/utils.tsx
@@ -1,10 +1,5 @@
-import { SelectOption } from 'components/activities/common/authoring/InputTypeDropdown';
 import { DEFAULT_PART_ID, setDifference, setUnion } from 'components/activities/common/utils';
-import {
-  MultiInput,
-  MultiInputSchema,
-  MultiInputType,
-} from 'components/activities/multi_input/schema';
+import { VlabInput, VlabSchema, VlabInputType } from 'components/activities/vlab/schema';
 import {
   makeChoice,
   makeHint,
@@ -22,12 +17,7 @@ import React from 'react';
 import { clone } from 'utils/common';
 import guid from 'utils/guid';
 
-export const multiInputOptions: SelectOption<'text' | 'numeric'>[] = [
-  { value: 'numeric', displayValue: 'Number' },
-  { value: 'text', displayValue: 'Text' },
-];
-
-export const multiInputStem = (input: InputRef) => ({
+export const vlabStem = (input: InputRef) => ({
   id: guid(),
   content: [
     {
@@ -38,15 +28,23 @@ export const multiInputStem = (input: InputRef) => ({
   ],
 });
 
-export const defaultModel = (): MultiInputSchema => {
+export const defaultModel = (): VlabSchema => {
   const input = Model.inputRef();
 
   return {
-    stem: multiInputStem(input),
+    stem: vlabStem(input),
     choices: [],
-    inputs: [{ inputType: 'text', id: input.id, partId: DEFAULT_PART_ID }],
+    inputs: [{ inputType: 'numeric', id: input.id, partId: DEFAULT_PART_ID }],
+    assignmentSource: 'builtIn',
+    assignmentPath: 'default',
+    assignment: DEFAULT_ASSIGNMENT,
+    configuration: DEFAULT_CONFIGURATION,
+    reactions: DEFAULT_REACTIONS,
+    solutions: DEFAULT_SOLUTIONS,
+    species: DEFAULT_SPECIES,
+    spectra: DEFAULT_SPECTRA,
     authoring: {
-      parts: [makePart(Responses.forTextInput(), [makeHint('')], DEFAULT_PART_ID)],
+      parts: [makePart(Responses.forNumericInput(), [makeHint('')], DEFAULT_PART_ID)],
       targeted: [],
       transformations: [makeTransformation('choices', Transform.shuffle)],
       previewText: 'Example question with a fill in the blank',
@@ -54,21 +52,46 @@ export const defaultModel = (): MultiInputSchema => {
   };
 };
 
-export const friendlyType = (type: MultiInputType) => {
-  if (type === 'dropdown') {
-    return 'Dropdown';
-  }
-  return `Input (${type === 'numeric' ? 'Number' : 'Text'})`;
+export const friendlyType = (type: VlabInputType) => {
+  return `${
+    type === 'dropdown'
+      ? 'Dropdown'
+      : type === 'vlabvalue'
+      ? 'Vlab Value'
+      : type === 'numeric'
+      ? 'Input (Number)'
+      : 'Input (Text)'
+  }`;
 };
 
-export const partTitle = (input: MultiInput, index: number) => (
+export const friendlyVlabParameter = (param: VlabParameter) => {
+  return `${
+    param === 'temp'
+      ? 'temperature (deg K)'
+      : param === 'volume'
+      ? 'volume (L)'
+      : param === 'moles'
+      ? 'moles'
+      : param === 'mass'
+      ? 'mass (g)'
+      : param === 'molarity'
+      ? 'molarity (moles/L)'
+      : param === 'concentration'
+      ? 'concentration (g/L)'
+      : param === 'pH'
+      ? 'pH'
+      : ''
+  }`;
+};
+
+export const partTitle = (input: VlabInput, index: number) => (
   <div>
     {`Part ${index + 1}: `}
     <span className="text-muted">{friendlyType(input.inputType)}</span>
   </div>
 );
 
-export function guaranteeMultiInputValidity(model: MultiInputSchema): MultiInputSchema {
+export function guaranteeMultiInputValidity(model: VlabSchema): VlabSchema {
   // Check whether model is valid first to save unnecessarily cloning the model
   if (isValidModel(model)) {
     return model;
@@ -80,7 +103,7 @@ export function guaranteeMultiInputValidity(model: MultiInputSchema): MultiInput
   );
 }
 
-function inputsMatchInputRefs(model: MultiInputSchema) {
+function inputsMatchInputRefs(model: VlabSchema) {
   const inputRefs = elementsOfType(model.stem.content, 'input_ref');
   const union = setUnion(
     inputRefs.map(({ id }) => id),
@@ -89,7 +112,7 @@ function inputsMatchInputRefs(model: MultiInputSchema) {
   return union.length === inputRefs.length && union.length === model.inputs.length;
 }
 
-function inputsMatchParts(model: MultiInputSchema) {
+function inputsMatchParts(model: VlabSchema) {
   const parts = model.authoring.parts;
   const union = setUnion(
     model.inputs.map(({ partId }) => partId),
@@ -98,7 +121,7 @@ function inputsMatchParts(model: MultiInputSchema) {
   return union.length === model.inputs.length && union.length === parts.length;
 }
 
-function inputsMatchChoices(model: MultiInputSchema) {
+function inputsMatchChoices(model: VlabSchema) {
   const inputChoiceIds = model.inputs.reduce(
     (acc, curr) => (curr.inputType === 'dropdown' ? acc.concat(curr.choiceIds) : acc),
     [] as string[],
@@ -110,11 +133,11 @@ function inputsMatchChoices(model: MultiInputSchema) {
   return union.length === model.choices.length && union.length === inputChoiceIds.length;
 }
 
-function hasAnInput(model: MultiInputSchema) {
+function hasAnInput(model: VlabSchema) {
   return model.inputs.length > 0;
 }
 
-function isValidModel(model: MultiInputSchema): boolean {
+function isValidModel(model: VlabSchema): boolean {
   return (
     hasAnInput(model) &&
     inputsMatchInputRefs(model) &&
@@ -123,7 +146,7 @@ function isValidModel(model: MultiInputSchema): boolean {
   );
 }
 
-function ensureHasInput(model: MultiInputSchema) {
+function ensureHasInput(model: VlabSchema) {
   if (hasAnInput(model)) {
     return model;
   }
@@ -143,10 +166,12 @@ function ensureHasInput(model: MultiInputSchema) {
   model.inputs.push(input);
   model.authoring.parts.push(part);
 
+  console.log('New model:' + model);
+
   return model;
 }
 
-function matchInputsToChoices(model: MultiInputSchema) {
+function matchInputsToChoices(model: VlabSchema) {
   if (inputsMatchChoices(model)) {
     return model;
   }
@@ -172,7 +197,7 @@ function matchInputsToChoices(model: MultiInputSchema) {
   return model;
 }
 
-function matchInputsToParts(model: MultiInputSchema) {
+function matchInputsToParts(model: VlabSchema) {
   if (inputsMatchParts(model)) {
     return model;
   }
@@ -188,7 +213,7 @@ function matchInputsToParts(model: MultiInputSchema) {
     model.authoring.parts.find((part) => part.id === id),
   );
 
-  unmatchedInputs.forEach((input: MultiInput) => {
+  unmatchedInputs.forEach((input: VlabInput) => {
     const choices = [makeChoice('Choice A'), makeChoice('Choice B')];
     const part = makePart(
       input.inputType === 'dropdown'
@@ -222,7 +247,7 @@ function matchInputsToParts(model: MultiInputSchema) {
   return model;
 }
 
-function matchInputsToInputRefs(model: MultiInputSchema) {
+function matchInputsToInputRefs(model: VlabSchema) {
   if (inputsMatchInputRefs(model)) {
     return model;
   }
@@ -238,7 +263,7 @@ function matchInputsToInputRefs(model: MultiInputSchema) {
     (id) => ({ id, type: 'input_ref' } as InputRef),
   );
 
-  unmatchedInputs.forEach((input: MultiInput) => {
+  unmatchedInputs.forEach((input: VlabInput) => {
     // add inputRef to end of first paragraph in stem
     const firstParagraph = model.stem.content.find((e) => e.type === 'p') as Paragraph | undefined;
     firstParagraph?.children.push({ ...Model.inputRef(), id: input.id });
@@ -253,3 +278,190 @@ function matchInputsToInputRefs(model: MultiInputSchema) {
   });
   return model;
 }
+
+const DEFAULT_ASSIGNMENT = '{ "assignmentText": "Assignment name" }';
+const DEFAULT_CONFIGURATION = `
+{
+  "title": "Stoichiometric Ratios",
+  "solutionViewers": [
+    {
+      "id": "solutionProperties",
+      "displayDefault": true,
+      "args": {
+        "honorSignificantFigures": false
+      }
+    },
+    {
+      "id": "aqueous",
+      "displayDefault": true,
+      "args": {
+        "unitsToggleEnabled": true
+      }
+    },
+    {
+      "id": "solid",
+      "displayDefault": false,
+      "args": {
+        "unitsToggleEnabled": true
+      }
+    },
+    {
+      "id": "spectrometer",
+      "displayDefault": false
+    },
+    {
+      "id": "particleView",
+      "displayDefault": true
+    },
+    {
+      "id": "thermometer",
+      "displayDefault": true
+    },
+    {
+      "id": "pH",
+      "displayDefault": true
+    },
+    {
+      "id": "vesselTrackingControl",
+      "displayDefault": false
+    }
+  ],
+  "transfer": ["precise", "significantFigures", "realistic"]
+}
+`;
+const DEFAULT_REACTIONS = `
+{
+  "REACTIONS": {
+    "REACTION": [
+      {
+        "SPECIES_REF": [
+          {
+            "id": "0", 
+            "coefficient": "1"
+          }, 
+          {
+            "id": "1", 
+            "coefficient": "-1"
+          }, 
+          {
+            "id": "2", 
+            "coefficient": "-1"
+          }
+        ]
+      }, 
+      {
+        "SPECIES_REF": [
+          {
+            "id": "5", 
+            "coefficient": "1"
+          }, 
+          {
+            "id": "3", 
+            "coefficient": "-1"
+          }, 
+          {
+            "id": "4", 
+            "coefficient": "-1"
+          }
+        ]
+      } 
+    ]
+  }
+}
+`;
+const DEFAULT_SOLUTIONS = `
+{
+  "FILESYSTEM": {
+    "DIRECTORY": [
+     {
+        "name": "stockroom", 
+        "SOLUTION": [
+          {
+            "name": "Distilled H<sub>2</sub>O", 
+            "description": "Distilled Water", 
+            "volume": "3.0", 
+            "vessel": "3LCarboy", 
+            "species": [
+               {
+                  "id": "0"
+               }
+            ]
+          },       
+          {
+            "name": "0.154 M NaCl", 
+            "description": "Sodium Chloride", 
+            "volume": "0.2", 
+            "species": [
+               {
+                  "id": "0"
+               }, 
+               {
+                  "id": "3", 
+                  "amount": "0.0308"
+               }, 
+               {
+                  "id": "4", 
+                  "amount": "0.0308"
+               }
+            ]
+          }
+        ]
+      }
+    ]
+  }
+}
+`;
+const DEFAULT_SPECIES = `
+{
+ "SPECIES_LIST": {
+  "SPECIES": [
+      {
+        "id": "0", 
+        "name": "H<sub>2</sub>O", 
+        "enthalpy": "-285.83", 
+        "entropy": "69.91", 
+        "state": "l", 
+        "molecularWeight": "18.016"
+      }, 
+      {
+        "id": "1", 
+        "name": "H<sup>+</sup>", 
+        "enthalpy": "0.0", 
+        "entropy": "0.0", 
+        "molecularWeight": "1.008"
+      }, 
+      {
+        "id": "2", 
+        "name": "OH<sup>-</sup>", 
+        "enthalpy": "-229.99", 
+        "entropy": "-10.75", 
+        "molecularWeight": "17.008"
+      }, 
+      {
+        "id": "3", 
+        "name": "Na<sup>+</sup>", 
+        "enthalpy": "-240.12", 
+        "entropy": "59.0", 
+        "molecularWeight": "22.99"
+      }, 
+      {
+        "id": "4", 
+        "name": "Cl<sup>-</sup>", 
+        "enthalpy": "-167.58999999999997", 
+        "entropy": "56.5", 
+        "molecularWeight": "35.45"
+      }, 
+      {
+        "id": "5", 
+        "name": "NaCl", 
+        "enthalpy": "-411.2", 
+        "entropy": "72.1", 
+        "state": "s", 
+        "molecularWeight": "58.44", 
+        "density": "2.16"
+     } 
+    ]
+  }
+}
+`;
+const DEFAULT_SPECTRA = '{ "SPECTRA_LIST": { "SPECIES": [] } }';